+++ ameriflux_pipeline/pyfluxpro/amerifluxformat.py
@@ -38,6 +38,8 @@ class AmeriFluxFormat:
         # remove meta data from dataframe
         full_output_df = full_output_df.iloc[1:, :]
         met_df = met_df.iloc[1:, :]
+        full_output_df.reset_index(drop=True, inplace=True)  # reset index after dropping rows
+        met_df.reset_index(drop=True, inplace=True)  # reset index after dropping rows
 
         # Step 1 of guide
         full_output_df = AmeriFluxFormat.replace_empty(full_output_df)
@@ -144,6 +146,7 @@ class AmeriFluxFormat:
             print("Albedo column not present")
             albedo_col = None
         if albedo_col:
+            met_df = met_df.astype({albedo_col: float})
             met_df['ALB'] = met_df[albedo_col].apply(lambda x: 1 if float(x) > 1 else float(x) * 100)
             met_df_meta['ALB'] = '%'
 
@@ -151,7 +154,7 @@ class AmeriFluxFormat:
             full_output_df['VPD'] = full_output_df['VPD'] / 100
             full_output_df_meta['VPD'].iloc[0] = '[hPa]'
         if 'Tau' in full_output_df:
-            full_output_df['Tau'] = full_output_df['Tau'].abs()
+            full_output_df['Tau'] = full_output_df['Tau'] * -1
             full_output_df_meta['Tau'].iloc[0] = '[kg+1m-1s-2]'
 
         # convert soil moisture variables into percentage values