+++ ameriflux_pipeline/utils/data_util.py
@@ -8,6 +8,7 @@ import pandas as pd
 import re
 import pathlib
 import os
+from dateutil.parser import parse
 
 
 def write_data(df, output_data):
@@ -20,7 +21,7 @@ def write_data(df, output_data):
         Returns:
             None
     """
-    print("Write data to file ", output_data)
+    print("Write data to csv file ", output_data)
     df.to_csv(output_data, index=False)
 
 
@@ -33,6 +34,7 @@ def read_excel(file_path):
         Returns:
             df (object): Pandas DataFrame object
     """
+    print("Read excel file ", file_path)
     df = pd.read_excel(file_path)  # read excel file
     return df
 
@@ -69,3 +71,19 @@ def get_directory(file_path):
     path = pathlib.Path(file_path)
     dir = os.path.dirname(path)
     return dir
+
+
+def get_valid_datetime(data):
+    """
+    Method to check if the string input date is in valid format recognizable by datetime
+    Returns a valid datetime if valid string input, else returns None
+    Args:
+        data (str): Input date to check for validity. The date can be in any format
+    Returns:
+        parse (str): Parsed datetime converted to string. None if data is not valid.
+    """
+    try:
+        return parse(data)
+    except ValueError:
+        print(data, "Incorrect date format")
+        return None